Add files via upload
LockedMe.com- Lockers pvt ltd. first project. It's a developed prototype of the application for the stakeholders of the company .
This project performs variety of action FILEs addition,deletion,search , in a specified directory.

Setup:
To run this project in your localhost - download this project in local directory and import it in your Eclipse IDE.